requirements.txt
telethon[cryptg]
aiohttp[speedups]
pyyaml